Data Cleaning in R description
As I mentioned in "Data Error Checking in Python", if we found a piece of data that only has numeric value but does not have units, what should we do about then? It's very hard to tell what unit should be add to the blank. Suppose you are given a piece of data of apple amount in 6,000. You never know whether it is supposed to be 6,000 tons of 6,000 kilograms. 

In this R code, I write a function called compareNearby which compares the unit missing county with other 5 to 10 nearby counties. My assumption is that all these counties share the same productivity. Therefore, if I see one nearby county that has the apple amount of 2,000 to 10,000 kilograms from 1949 to 1990, then I know the unit missing county with apple amount of 6,000 should have 6,000 kilograms apple. After adding all the missing unit in the blank by compareNearby function, I unify all the units from "pounds", "kilograms", "tons" to "kilograms" and convert the data to "kilogram" magnitude accordingly. Last, I called the compareNearby function again and this time I output the scatter plot of the unit missing county and other 5 nearby counties. If I am right about adding units to the unit missing counties, I should see the same data magnitude of all the 6 counties. If not, I should go back to the original data to further check if I am wrong about adding units.